Extract shared DatasetBrowser component to eliminate duplication
Both /data (tab) and /data/browser (standalone page) rendered identical
dataset browser code. Extracted into components/data/DatasetBrowser.tsx
so both routes import a single source of truth.
